added deleting; fixed bugs; added vertice drawing

+++ vertice.cpp
@@ -1,7 +1,7 @@
 #include "vertice.h"
 
-uint8_t Vertice::x() { return x_; }
-uint8_t Vertice::y() { return y_; }
+int Vertice::x() { return x_; }
+int Vertice::y() { return y_; }
 
-void Vertice::SetX(uint8_t&& x) { x_ = x; }
-void Vertice::SetY(uint8_t&& y) { y_ = y; }
+void Vertice::SetX(int&& x) { x_ = x; }
+void Vertice::SetY(int&& y) { y_ = y; }
